Fix Linux/GCC build breaks, unbreak `git clone --recursive`, add CI

+++ src/RefCounter/refcounter.c
@@ -20,7 +20,7 @@ void refcounter_yield(refcounter_t* refcounter) {
 void* refcounter_reference(refcounter_t* refcounter) {
     if (refcounter == NULL) return NULL;
     // Try to consume a yield first, then fall back to incrementing count
-    uint8_t expected_yield = atomic_load(&refcounter->yield);
+    uint_fast8_t expected_yield = atomic_load(&refcounter->yield);
     while (expected_yield > 0) {
         if (atomic_compare_exchange_weak(&refcounter->yield, &expected_yield, expected_yield - 1)) {
             return refcounter;
@@ -35,7 +35,7 @@ void* refcounter_reference(refcounter_t* refcounter) {
 void refcounter_dereference(refcounter_t* refcounter) {
     if (refcounter == NULL) return;
     // Try to consume a yield first (matching refcounter_reference pattern)
-    uint8_t expected_yield = atomic_load(&refcounter->yield);
+    uint_fast8_t expected_yield = atomic_load(&refcounter->yield);
     while (expected_yield > 0) {
         if (atomic_compare_exchange_weak(&refcounter->yield, &expected_yield, expected_yield - 1)) {
             return;  // Consumed a yield, don't decrement count
@@ -47,9 +47,9 @@ void refcounter_dereference(refcounter_t* refcounter) {
     // >0 check, and both decrement — underflowing the count to 65535.
     // This caused use-after-free → double-free chains in the memory pool.
     // The CAS loop ensures the count never goes below 0.
-    uint16_t expected = (uint16_t)atomic_load(&refcounter->count);
+    uint_fast16_t expected = atomic_load(&refcounter->count);
     while (expected > 0) {
-        if (atomic_compare_exchange_weak(&refcounter->count, &expected, (uint16_t)(expected - 1))) {
+        if (atomic_compare_exchange_weak(&refcounter->count, &expected, (uint_fast16_t)(expected - 1))) {
             return;  // Successfully decremented
         }
         // CAS failed — expected was updated with the current value, retry
